few revision

+++ data_utils.py
@@ -144,5 +144,5 @@ def dataset_resize(root,resize_dst=(512, 512)):
     for img_path in imgs_list:
         img = cv2.imread(img_path)
         img = img2square(img, img_size=resize_dst[0])
-        Image.fromarray(img).save(img_path)
+        Image.fromarray(img.astype(np.uint8)).save(img_path)
         print(img_path, 'is resized.')
